refactor: common logic was moved to GenerationStrategy
- GenerationStrategy now is an abstract class instead of interface with abstract property objectMapper
- SchemaFileSettings.format was unmarked as deprecated
