+++ src/main/kotlin/com/asyncapi/plugin/core/generator/GenerationStrategy.kt
@@ -1,7 +1,10 @@
 package com.asyncapi.plugin.core.generator
 
 import com.asyncapi.plugin.core.generator.exception.AsyncAPISchemaGenerationException
-import kotlin.jvm.Throws
+import com.asyncapi.plugin.core.generator.settings.GenerationSettings
+import com.asyncapi.plugin.core.io.AsyncAPISchemaLoader
+import com.asyncapi.plugin.core.io.FileSystem
+import com.fasterxml.jackson.databind.ObjectMapper
 
 /**
  * AsyncAPI schema generation strategy.
@@ -9,7 +12,13 @@ import kotlin.jvm.Throws
  * @author Pavel Bodiachevskii
  * @since 1.0.0-RC1
  */
-interface GenerationStrategy {
+abstract class GenerationStrategy(
+    private val generationSettings: GenerationSettings
+) {
+
+    private val asyncAPISchemaLoader = AsyncAPISchemaLoader(generationSettings.logger, generationSettings.sources)
+
+    abstract val objectMapper: ObjectMapper
 
     /**
      * Generates AsyncAPI Schema.
@@ -17,6 +26,33 @@ interface GenerationStrategy {
      * @throws AsyncAPISchemaGenerationException in case of schema generation issues.
      */
     @Throws(AsyncAPISchemaGenerationException::class)
-    fun generate()
+    fun generate() {
+        val schemas = asyncAPISchemaLoader.load()
+        schemas.forEach(this::save)
+    }
+
+    private fun save(schemaClass: Class<*>) {
+        val schema = serialize(schemaClass)
+        val schemaFileName = "${schemaClass.simpleName}-${generationSettings.schemaFile.namePostfix}.${generationSettings.schemaFile.format}"
+        val schemaFilePath = generationSettings.schemaFile.path
+
+        generationSettings.logger.info("saving ${schemaClass.name} to $schemaFilePath")
+        FileSystem.save(schemaFileName, schema, schemaFilePath)
+    }
+
+    @Throws(AsyncAPISchemaGenerationException::class)
+    private fun serialize(schemaClass: Class<*>): String {
+        return try {
+            val foundAsyncAPI = schemaClass.newInstance()
+
+            if (generationSettings.rules.prettyPrint) {
+                objectMapper.writerWithDefaultPrettyPrinter().writeValueAsString(foundAsyncAPI)
+            } else {
+                objectMapper.writeValueAsString(foundAsyncAPI)
+            }
+        } catch (exception: Exception) {
+            throw AsyncAPISchemaGenerationException("Can't serialize: ${schemaClass.simpleName}. Because of: ${exception.message}", exception)
+        }
+    }
 
 }
